+++ lib/src/features/super_table/presentation/pages/super_table_demo.dart
@@ -11,7 +11,7 @@
 
 import 'package:flutter/material.dart';
 
-import '../../../../core/core.dart';
+import 'package:super_auto_suggestion_box/super_auto_suggestion_box.dart';
 import '../../domain/entities/super_column.dart';
 import '../../domain/entities/super_columns.dart';
 import '../../domain/entities/super_row.dart';